SAPRINCore

This module creates the standard SAPRIN Episode datasets

Reference

SAPRINCore.addindividualattributes — Method

Add individual characteristics to to day records

SAPRINCore.addresidentdayswithoutmembership — Method

Consolidate daily exposure for an individual to a single preferred household based on relationship to household head

SAPRINCore.basicepisodes — Method

Group day records into basic exposure records

SAPRINCore.batchmembershipdays — Method

Extract membership days in batches

SAPRINCore.batchmemberships — Function

Normalise memberships in batches

SAPRINCore.batchonindividualid — Function

Convert a Arrow file into an Arrow file with record batches based on IndividualId

SAPRINCore.batchresidencydays — Method

Normalise residencies in batches

SAPRINCore.bornflag — Method

Determine whether the born flag should be set

SAPRINCore.combinebatches — Method

Concatenate record batches

SAPRINCore.combinemembershipbatch — Method

Concatenate membership batches

SAPRINCore.createdirectories — Method

Create directories to output staging data

SAPRINCore.diedflag — Method

Determine whether the died end flag should be set

SAPRINCore.dropnonresidentepisodes — Method

Use resident status observations to identify non-resident episodes and drop those from residency episodes

SAPRINCore.eliminateresidenceoverlaps — Method

Decompose residences into days and eliminate overlaps

SAPRINCore.enumerationflag — Method

Determine whether the enumeration flag should be set

SAPRINCore.extresstartflg — Method

Determine whether the external residency start flag should be set

SAPRINCore.individualobservationbounds — Method

Creates a dataset with the earliest and latest date at which an individual has been observed within left and rightcensor dates

SAPRINCore.inmigrationflag — Method

Determine whether the inmigration flag should be set

SAPRINCore.locationentryflag — Method

Determine whether the participation start flag should be set

SAPRINCore.ltfuflag — Method

Determine whether the lost to follow-up end flag should be set

SAPRINCore.newrandomdate — Method

Returns a random date relative to base, constraint by start and end date

SAPRINCore.normalisehouseholdresidencies — Method

Ensure household residency span household membership extends

SAPRINCore.participationflag — Method

Determine whether the participation start flag should be set

SAPRINCore.processresidencedays — Method

Decompose residency episodes into days and eliminate overlaps

SAPRINCore.processresidencydays — Method

Extract household residency days

SAPRINCore.readhouseholdheadrelationships — Method

Retrieve household head relationships with left and right censor dates

SAPRINCore.readhouseholdmemberships — Method

Retrieve and save household membership and householdhead relationships data

SAPRINCore.readhouseholdmemberships_internal — Method

Retrieve household memberships with left and right censor dates

SAPRINCore.readhouseholdresidences — Method

Retrieve household residencies with left and right censor dates

SAPRINCore.readhouseholds — Method

Read households and save household map and data to staging directory

SAPRINCore.readhouseholds_internal — Method

Read households and save household map and data to staging directory

SAPRINCore.readhouseholdsocioeconomic — Method

Read and save household socio-economic data

SAPRINCore.readindividualmemberships — Function

Decompose household memberships to days and combine with houeshold relationships and produced consolidated individual hosuehold membership episodes with household head relationship

SAPRINCore.readindividuals — Method

Read individuals and anonimise for specified node specified and save individual data, id map and bounds to to arrow files

SAPRINCore.readindividuals_internal — Method

Read individuals and anonimise for node specified in settings and save id map and individual data to to arrow files

SAPRINCore.readlocations_internal — Method

Read locations and anonimise for node save id map and location data to to arrow files

SAPRINCore.readmaritalstatuses — Method

Read and save individual marital status observations

SAPRINCore.readresidences — Method

Retrieve and save residence episodes, assumes individual and locations have been read

SAPRINCore.readresidences_internal — Method

Retrieve and save residence episodes directly from database

SAPRINCore.readresidencestatus — Method

Read resident status observations

SAPRINCore.recoverresidentdays — Method

Recreate membership days for residency days without a membership

SAPRINCore.refusalflag — Method

Determine whether the refusal end flag should be set

SAPRINCore.rightcensor — Method

Constrain date a to be no larger than b

SAPRINCore.setdayflags! — Method

Set residency start and end flags for a day

SAPRINCore.setresidencyflags — Method

Determine whether a particular day starts or end a residency episode and the type of start or end